Method-Overhaul-&-Renaming
Overhauled the methods implementing options in terms of dealing with errors.
You can now either chose for the method to auto retry broken hyperlinks or for it to finish without and print a log containing the problematic urls) for both the gather and the crafter methods.

Additionally added error logging to the Garland fetcher method.

Furthermore added toggles on the ItemSorter methods that allow for switching between duplicity removal and 1:1 mode. (Defaults to duplicity removal)

Also properly Identified methods that are internal as internal. (you should not need to interact with these unless your adding to the class itself)

NOTE: during this push the test functions/example code were commented out and left as an example.